Feat: added task status reordering functionality

+++ internal/repository/model/project_task_type.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type ProjectTaskType struct {
-	Id        int       `gorm:"column:id;primary_key;AUTO_INCREMENT" json:"id"`
+	Id        int64     `gorm:"column:id;primary_key;AUTO_INCREMENT" json:"id"`
 	Title     string    `gorm:"column:title" json:"title"`
 	CreatedBy int       `gorm:"column:created_by" json:"created_by"`
 	CreatedAt time.Time `gorm:"column:created_at" json:"created_at"`
